Gatling Enemy/ Astroid/ Enemy Health Bar's/ Bullet and target Tracking
Planning on implementing different firing modes such as burst fire and ect
but its to late as of now. Still waiting for assets and animations from designer. Also made the health bar accurate to the ships health and damage taken from the weapon script. its set to 100 though so may need to modify code a bit to make it like enemy health bars which auto adjust